Split the site into separate pages.
