+++ .postcssrc
@@ -1,5 +1,5 @@
 {
-  "plugins": {
-    "tailwindcss": {}
-  }
+	"plugins": {
+		"tailwindcss": {}
+	}
 }
