+++ .posthtmlrc
@@ -1,0 +1,7 @@
+{
+	"plugins": {
+		"posthtml-include": {
+			"root": "./src"
+		}
+	}
+}
