+++ src/header.html
@@ -1,14 +1,14 @@
 <header class="flex justify-between">
-	<div class="p-6">
-		<img src="img/logo.svg" alt="Media over QUIC" />
+	<div>
+		<img src="common/logo.svg" class="h-32" alt="Media over QUIC" />
 	</div>
 	<nav class="flex gap-12 p-6">
-		<a href="#">Watch</a>
-		<a href="#">Broadcast</a>
-		<a href="#">About</a>
+		<a href="watch.html">Watch</a>
+		<a href="publish.html">Publish</a>
+		<a href="explained.html">Explained</a>
 	</nav>
 	<div>
-		<a href="https://github.com/kixelated/moq-rs" class="github-corner">
+		<a href="https://github.com/kixelated/quic.video" class="github-corner">
 			<svg width="80" height="80" viewBox="0 0 250 250" style="fill: #151513; color: #fff" aria-hidden="true">
 				<path d="M0,0 L115,115 L130,115 L142,142 L250,250 L250,0 Z"></path>
 				<path
