+++ src/publish.html
@@ -1,0 +1,24 @@
+﻿<!doctype html>
+<html>
+	<head>
+		<meta charset="UTF-8" />
+		<title>quic.video</title>
+
+		<link rel="stylesheet" href="common/base.css" />
+	</head>
+
+	<body>
+		<include src="header.html"></include>
+		<div id="main" class="relative flex-grow p-6"></div>
+
+		<script type="module">
+			import { Main } from "./publish/index.tsx"
+			import { render } from "solid-js/web"
+
+			const main = document.getElementById("main")
+			if (!main) throw new Error("no container")
+
+			render(() => <Main />, main)
+		</script>
+	</body>
+</html>
